src/api/requirements.txt
fastapi
uvicorn
pydantic
boto3
asyncio
sqlalchemy
python-dotenv
gunicorn
websockets
aioredis
pillow
requests
tenacity
setuptools

# These are necessary to run the intent classifier
# TODO: Consolidate requirements files
Jinja2==3.1.4
langchain==0.2.16
langchain-anthropic==0.1.22
langchain-community==0.2.16
langchain-core==0.2.39
langchain-openai==0.1.23
langchain-text-splitters==0.2.2
langfuse==2.33.0
